+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/menu/energy_transfer_node/EnergyTransferNodeMenu.kt
@@ -1,11 +1,9 @@
 package dev.aaronhowser.mods.excessive_utilities.menu.energy_transfer_node
 
 import dev.aaronhowser.mods.aaron.menu.MenuWithInventory
-import dev.aaronhowser.mods.aaron.menu.components.FilteredSlot
-import dev.aaronhowser.mods.aaron.misc.AaronExtensions.isItem
 import dev.aaronhowser.mods.excessive_utilities.block_entity.base.TransferNodeBlockEntity
 import dev.aaronhowser.mods.excessive_utilities.block_entity.transfer_node.EnergyTransferNodeBlockEntity
-import dev.aaronhowser.mods.excessive_utilities.datagen.tag.ModItemTagsProvider
+import dev.aaronhowser.mods.excessive_utilities.menu.TransferNodeUpgradeSlot
 import dev.aaronhowser.mods.excessive_utilities.registry.ModMenuTypes
 import net.minecraft.world.Container
 import net.minecraft.world.SimpleContainer
@@ -41,9 +39,7 @@ class EnergyTransferNodeMenu(
 			val x = 35 + i * 18
 			val y = 68
 
-			val slot = FilteredSlot(upgradesContainer, i, x, y) {
-				it.isItem(ModItemTagsProvider.TRANSFER_NODE_UPGRADES) || it.isItem(ModItemTagsProvider.RETRIEVAL_NODE_UPGRADES)
-			}
+			val slot = TransferNodeUpgradeSlot(upgradesContainer, i, x, y)
 
 			this.addSlot(slot)
 		}